requirements.txt
streamlit~=1.45.1
psycopg2-binary~=2.9.10
pandas~=2.2.3

bcrypt~=4.3.0
python-dotenv~=1.1.0
yagmail~=0.15.293
reportlab~=4.4.1
arabic-reshaper~=3.0.0
python-bidi~=0.6.6
matplotlib~=3.10.3
plotly~=6.1.1
openai~=1.82.0
pillow~=11.2.1
tornado==6.4.2

supabase~=2.15.2
PyDrive~=1.3.1
PyPDF2~=3.0.1
pdf2image~=1.17.0
google-cloud-storage~=2.16.0
